+++ changes.md
@@ -1,4 +1,6 @@
 # Recent changes
+## 3 July 2023
+- Moved delete style functionality into an API call
 
 ## 1 July 2023
 - Moved `create new additional style file` to the dropdown menu